test performance.now as well. Refactor
