--- tests/compiler/wasi/date.ts
@@ -1,5 +1,0 @@
-import "wasi";
-
-export function test(): i64 {
-  return Date.now();
-}
