+++ tests/compiler/wasi/now.json
